Moved react components into a components sub directory
