+++ client/components/badge.js
